Refactored the new <Sheet> which was on multiple places and created <WindowSheet> which collects all (most of) the custom stuff.
